ui(cli): modularize the interface spec by splitting files and organizing
The interface file was both without a clear goal, and it was getting far too
long to be manageable in a typical fashion. As such, by creating a new submodule
for the interface components, we can reorganize the groupings into things that
  make more sense to me. Plus we can add a bar owl to the mix...
